Fix invalid Postgis and PostgreSQL formulae
When attempting to access the `vibes/software` tap, Homebrew complained
and returned a non-zero error code, stating that the formulae for
Postgis and PostgreSQL were invalid - thus, not allowing for software to
be installed via the tap.

This commit runs the suggested `brew style --fix` command against the
two offending files:

    * `postgis.rb`
    * `postgresql.rb`

This makes the necessary changes and, along with some additional
stylistic changes, should allow devs to successfully tap and install
software again.